Refactor AppIntro fragments into new package
