+++ app/src/main/java/com/lambdaschool/empoweredconversation/fragment/intro/IntroSlide1Fragment.kt
@@ -1,17 +1,17 @@
-package com.lambdaschool.empoweredconversation.fragment
+package com.lambdaschool.empoweredconversation.fragment.intro
 
 import android.os.Bundle
-import androidx.fragment.app.Fragment
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
-
+import androidx.fragment.app.Fragment
 import com.lambdaschool.empoweredconversation.R
 
-class IntroSlide2Fragment : Fragment() {
+class IntroSlide1Fragment : Fragment() {
 
     override fun onCreateView(inflater: LayoutInflater, container: ViewGroup?, savedInstanceState: Bundle?): View? {
-        return inflater.inflate(R.layout.fragment_intro_slide2, container, false)
+        return inflater.inflate(R.layout.fragment_intro_slide1, container, false)
     }
+
 
 }
