+++ app/src/main/res/layout/fragment_intro_slide3.xml
@@ -3,7 +3,7 @@
                                                    xmlns:app="http://schemas.android.com/apk/res-auto" xmlns:tools="http://schemas.android.com/tools"
                                                    android:layout_width="match_parent"
                                                    android:layout_height="match_parent"
-                                                   tools:context=".fragment.IntroSlide1Fragment"
+                                                   tools:context=".fragment.intro.IntroSlide1Fragment"
                                                    android:background="@color/colorPrimary">
     <androidx.constraintlayout.widget.ConstraintLayout
             android:layout_width="match_parent"
